Profile settings (#81)
* small tweaks

* added noAuth prop to next pages

* pfp component, browser util

* header search bar

* mmm

* most of ui done

* done :)

+++ src/server/uploadthing.ts
@@ -1,8 +1,10 @@
+import { userRepository } from "@/database/repository/auth";
+import { reviewRepository } from "@/database/repository/product";
+import { isValidUrlString } from "@/utils";
 import { createUploadthing, type FileRouter } from "uploadthing/next-legacy";
+import { utapi } from "uploadthing/server";
 import { z } from "zod";
 import { getServerAuthSession } from "./auth";
-import { reviewRepository } from "@/database/repository/product";
-import { utapi } from "uploadthing/server";
 
 const uploadthing = createUploadthing();
 
@@ -31,9 +33,37 @@ export const appFileRouter = {
           and(eq(table.userId, userId), eq(table.barcode, barcode))
         )
         .then((query) => {
-          if (!(query.rowsAffected && oldImageKey)) return;
+          if (!query.rowsAffected || !oldImageKey) return;
 
           utapi.deleteFiles(oldImageKey).catch(console.error);
+        })
+        .catch((err) => {
+          console.error(err);
+          utapi.deleteFiles(file.key).catch(console.error);
+        });
+    }),
+  userImageUploader: uploadthing({ image: { maxFileSize: "512KB", maxFileCount: 1 } })
+    .middleware(async ({ req, res }) => {
+      const session = await getServerAuthSession({ req, res });
+      if (!session) throw Error("Unauthorized");
+
+      const user = await userRepository.findFirst((table, { eq }) => eq(table.id, session.user.id));
+      if (!user) throw Error("User not found");
+
+      return {
+        userId: session.user.id,
+        userImageKey: user.image && !isValidUrlString(user.image) ? user.image : null,
+      };
+    })
+    .onUploadError(({ error }) => {
+      console.log(error);
+    })
+    .onUploadComplete(({ file, metadata: { userId, userImageKey } }) => {
+      userRepository
+        .update({ image: file.key }, (table, { eq }) => eq(table.id, userId))
+        .then((query) => {
+          if (!query.rowsAffected || !userImageKey) return;
+          utapi.deleteFiles(userImageKey).catch(console.error);
         })
         .catch((err) => {
           console.error(err);
